+++ src/constants/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    GAS_FEE_API_ETH: 'https://gas-api.metaswap.codefi.network/networks/#{chainid}/suggestedGasFees'
+}